Made Action Bar work! Added BaseActivity class.
