+++ app/src/main/java/edu/calvin/cs262/climbinglogapp/BaseActivity.java
@@ -1,0 +1,57 @@
+package edu.calvin.cs262.climbinglogapp;
+
+import android.app.AlertDialog;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v4.app.FragmentActivity;
+import android.view.Menu;
+import android.view.MenuItem;
+
+/**
+ * Created by Austin on 11/8/2015.
+ * This base activity will be extended by every other class that needs the action bar.
+ */
+public class BaseActivity extends FragmentActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        //enable the action bar home button
+        getActionBar().setHomeButtonEnabled(true);
+    }
+    //Creating the options menu on the Action Bar
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        return true;
+    }
+
+    //Action Bar options method
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle presses on the action bar items
+        switch (item.getItemId()) {
+            //handle the action bar home button press
+            case android.R.id.home:
+                Intent mainIntent = new Intent(this, MainActivity.class);
+                startActivity(mainIntent);
+                return true;
+            case R.id.action_help:
+                //simple dialog when the help setting is selected
+                AlertDialog alertDialog = new AlertDialog.Builder(BaseActivity.this).create();
+                alertDialog.setTitle(getString(R.string.action_help));
+                alertDialog.setMessage(getString(R.string.action_help_content));
+                alertDialog.setButton(AlertDialog.BUTTON_NEUTRAL, "OK",
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog, int which) {
+                                dialog.dismiss();
+                            }
+                        });
+                alertDialog.show();
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+}
