+++ app/src/main/java/edu/calvin/cs262/climbinglogapp/Friends.java
@@ -1,16 +1,13 @@
 package edu.calvin.cs262.climbinglogapp;
 
-import android.support.v4.app.FragmentActivity;
 import android.os.Bundle;
 import android.widget.Button;
-
-import java.nio.BufferUnderflowException;
 
 /**
  * Created by jbu2 on 11/4/2015.
  * Friends page
  */
-public class Friends extends FragmentActivity {
+public class Friends extends BaseActivity {
 
     //onCreate() method
     protected void onCreate(Bundle savedInstanceState) {
