+++ app/src/main/java/edu/calvin/cs262/climbinglogapp/Profile.java
@@ -1,14 +1,16 @@
 package edu.calvin.cs262.climbinglogapp;
 
-import android.support.v4.app.FragmentActivity;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.os.Bundle;
+import android.view.MenuItem;
 import android.widget.Button;
 
 /**
  * Created by cpd5 on 10/12/2015.
  * This is the page that will contain the profile of the user.
  */
-public class Profile extends FragmentActivity {
+public class Profile extends BaseActivity {
 
     //onCreate() method
     protected void onCreate(Bundle savedInstanceState) {
@@ -16,6 +18,32 @@ public class Profile extends FragmentActivity {
         setContentView(R.layout.profile);
         Button disableProfile = (Button) findViewById(R.id.profile_button);  //Disable the corresponding Profile button
         disableProfile.setEnabled(false);  //To keep people from creating the Profile activity over and over again
+    }
+
+    /**
+     * This method overrides the onOptionsItemSelected method in order to handle action bar things
+     * specific to this page (eg the help dialog).
+     */
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle presses on the action bar items
+        switch (item.getItemId()) {
+            case R.id.action_help:
+                //simple dialog when the help setting is selected
+                AlertDialog alertDialog = new AlertDialog.Builder(Profile.this).create();
+                alertDialog.setTitle(getString(R.string.action_help));
+                alertDialog.setMessage(getString(R.string.profile_label));
+                alertDialog.setButton(AlertDialog.BUTTON_NEUTRAL, "OK",
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog, int which) {
+                                dialog.dismiss();
+                            }
+                        });
+                alertDialog.show();
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
     }
     //onPause method
     protected void onPause() {
